# Improve ModuleInstance String() performance