--- a/addrs/module_instance_test.go
+++ b/addrs/module_instance_test.go
@@ -77,3 +77,17 @@ func TestModuleInstanceEqual_false(t *testing.T) {
 		})
 	}
 }
+
+func BenchmarkStringShort(b *testing.B) {
+        addr, _ := ParseModuleInstanceStr(`module.foo`)
+        for n := 0; n < b.N; n++ {
+                addr.String()
+        }
+}
+
+func BenchmarkStringLong(b *testing.B) {
+        addr, _ := ParseModuleInstanceStr(`module.southamerica-brazil-region.module.user-regional-desktops.module.user-name`)
+        for n := 0; n < b.N; n++ {
+                addr.String()
+        }
+}