`bytes.Buffer` always allocates when converting to a string. `strings.Builder` does not.

Preallocating the memory is beneficial as well, note that the new code only does a single allocation per operation

Results, first this branch, then main:

```
$ go test -bench=String ./addrs -benchmem 
BenchmarkStringShort-12         18271692                56.52 ns/op           16 B/op          1 allocs/op
BenchmarkStringLong-12           8057071               158.5 ns/op            96 B/op          1 allocs/op
PASS
$ git checkout main addrs/module_instance.go
$ go test -bench=String ./addrs -benchmem 
BenchmarkStringShort-12          7690818               162.0 ns/op            80 B/op          2 allocs/op
BenchmarkStringLong-12           2922117               414.1 ns/op           288 B/op          3 allocs/op
```

This improves running `terraform validate` of some large configuration by ~15% of wall time